widget/tests/test_bug485118.xhtml

<?xml version="1.0"?>
<?xml-stylesheet href="chrome://global/skin" type="text/css"?>
<?xml-stylesheet href="chrome://mochikit/content/tests/SimpleTest/test.css"
                 type="text/css"?>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=485118
-->
<window title="Mozilla Bug 485118"
  xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul">

  <script src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js" />

<body  xmlns="http://www.w3.org/1999/xhtml">
<div id="content" style="display: none">

</div>
<pre id="test">
</pre>
</body>

<hbox height="300">
  <vbox width="300">
    <scrollbar id="horizontal"/>
    <hbox flex="1">
      <scrollbar vertical="true" id="vertical"/>
      <spacer flex="1"/>
    </hbox>
  </vbox>
</hbox>

<script class="testbody" type="application/javascript">
<![CDATA[

SimpleTest.waitForExplicitFinish();

function runTest() {
  ["horizontal", "vertical"].forEach(function (orient) {
    var elem = document.getElementById(orient);
    var thumbRect = SpecialPowers.unwrap(
      SpecialPowers.InspectorUtils.getChildrenForNode(elem, true, false)[0])
      .childNodes[0].getBoundingClientRect();
    var sizeToCheck = orient == "horizontal" ? "width" : "height";
    var expectedSize = 26;
    is(thumbRect[sizeToCheck], expectedSize, "scrollbar has wrong minimum " + sizeToCheck);
  });
  SimpleTest.finish();
}
window.addEventListener("load", runTest);

]]>
</script>

</window>
